fix(engines): containment helper moves to a leaf module (CodeQL cyclic-import)
utils/containment is stdlib-only, so model_manager and asr_backend both
import it at module top with no cycle — the call-time back-import CodeQL
flagged is gone.
